refactor: adder 클래스 함수 분리 및 int보다 큰 값 처리 구현

+++ src/main/java/calculator/Application.java
@@ -2,6 +2,7 @@ package calculator;
 
 import calculator.adder.Adder;
 import calculator.inputhandler.InputHandler;
+import java.math.BigInteger;
 
 public class Application {
     public static void main(String[] args) {
@@ -9,7 +10,7 @@ public class Application {
         String input = inputHandler.getInput("덧셈할 문자열을 입력해주세요.\n");
 
         Adder adder = new Adder();
-        int sum = adder.addInput(input);
+        BigInteger sum = adder.addInput(input);
     }
 
 }
